Accept scalar value for linvel instead
This will replace the stand-in constant DEATH_VEL which was used along
with the angle a specific particle was from the center to calculate a
proper velocity vector.

An additional field was added called linvel_addtl that will be added
flatly to the end velocity of the above mentioned calculated. Mainly to
be used for adding more noise and chaos to the particles.